add form component
